requirements.txt
supabase>=2.23.0
python-dotenv==1.0.0
requests==2.31.0
websockets>=15.0.0

